The markup of the newsletter subscription form and all its elements in the footer has been completed.
The markup of the application form and all its elements in the modal window has been completed.
All inputs in forms have a name attribute.

The values for the name attribute are descriptive, describing exactly what the form field is for.
All inputs have an associated <label> element.
The placeholder attribute is set for inputs if there is a hint text for it in the layout.
The type = "submit" attribute is set for the "submit" buttons of the forms.